backend/api/requirements.txt
# AgriSmart unified API (disease + IFS + history)
fastapi[standard]>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0

# Disease model (TensorFlow)
tensorflow>=2.16.0; sys_platform != 'darwin' or platform_machine != 'arm64'
tensorflow-macos>=2.16.0; sys_platform == 'darwin' and platform_machine == 'arm64'
tf-keras>=2.16.0
pillow>=10.0.0
numpy>=1.24.0,<2.0.0
